move functions to seperate files, wrap osm api calls into seperate try catch block, rename error/success handlers
