--- src/lib/fetchers/osm-api/makeChangeRequestToOsmApi.ts
@@ -1,217 +1,0 @@
-import React from "react";
-import type { SWRResponse } from "swr";
-import { log } from "../../util/logger";
-import { callBackendToUpdateInhouseDb } from "../callBackendToUpdateInhouseDb";
-import useInhouseOSMAPI from "./useInhouseOSMAPI";
-
-export async function createChangeset({
-  baseUrl,
-  tagName,
-  newValue,
-  accessToken,
-}: {
-  baseUrl: string;
-  tagName: string;
-  newValue: string;
-  accessToken: string;
-}): Promise<string> {
-  // Comments cannot be longer than 255 chars
-  const truncatedNewValue = newValue.substring(0, 100);
-  const response = await fetch(`${baseUrl}/api/0.6/changeset/create`, {
-    method: "PUT",
-    headers: {
-      "Content-Type": "text/xml; charset=UTF-8",
-      Authorization: `Bearer ${accessToken}`,
-    },
-
-    body: `<osm>
-      <changeset>
-        <tag k="created_by" v="https://wheelmap.org" />
-        <tag k="comment" v="Change ${tagName} tag to '${truncatedNewValue}'" />
-      </changeset>
-    </osm>`,
-  });
-  return response.text();
-}
-
-export type OSMAPIElement = {
-  version: string;
-  lat: string;
-  lon: string;
-  id: string;
-  tags: Record<string, string>;
-};
-export type OSMAPIElementResponse = SWRResponse<OSMAPIElement>;
-
-export async function createChange({
-  accessToken,
-  baseUrl,
-  osmType,
-  osmId,
-  changesetId,
-  tagName,
-  newTagValue,
-  currentOSMObjectOnServer,
-}: {
-  baseUrl: string;
-  accessToken: string;
-  osmType: string;
-  osmId: number;
-  changesetId: string;
-  tagName: string;
-  newTagValue: string;
-  currentOSMObjectOnServer: OSMAPIElement;
-}) {
-  log.log(
-    "createChange",
-    osmType,
-    osmId,
-    changesetId,
-    tagName,
-    newTagValue,
-    currentOSMObjectOnServer,
-  );
-  const { version, lat, lon, id, tags } = currentOSMObjectOnServer;
-  const numericId = id;
-  const currentTagsOnServer = tags;
-  const newTags = {
-    ...currentTagsOnServer,
-    [tagName]: newTagValue,
-  };
-
-  function escapeXML(value: string) {
-    return value
-      .replace(/&/g, "&amp;")
-      .replace(/</g, "&lt;")
-      .replace(/>/g, "&gt;")
-      .replace(/"/g, "&quot;")
-      .replace(/'/g, "&apos;");
-  }
-
-  const allTagsAsXML = Object.entries(newTags)
-    .map(
-      ([key, value]) => `<tag k="${escapeXML(key)}" v="${escapeXML(value)}" />`,
-    )
-    .join("\n");
-
-  log.log("allTagsAsXML", allTagsAsXML);
-
-  let body: string;
-
-  if (osmType === "node") {
-    body = `<osm>
-      <${osmType} id="${numericId}" changeset="${changesetId}" lat="${lat}" lon="${lon}" version="${version}">
-        ${allTagsAsXML}
-      </${osmType}>
-    </osm>`;
-  } else {
-    // for way and relation exclude lat and lon
-    body = `<osm>
-      <${osmType} id="${numericId}" changeset="${changesetId}" version="${version}">
-        ${allTagsAsXML}
-      </${osmType}>
-    </osm>`;
-  }
-  return fetch(`${baseUrl}/api/0.6/${osmId}`, {
-    method: "PUT",
-    headers: {
-      "Content-Type": "text/xml",
-      Authorization: `Bearer ${accessToken}`,
-    },
-    body,
-  })
-    .then((res) => res.text())
-    .then((data) => {
-      log.log(data);
-    });
-}
-
-export default function useSubmitNewValueCallback({
-  accessToken,
-  baseUrl,
-  osmType,
-  osmId,
-  tagName,
-  newTagValue,
-  currentOSMObjectOnServer,
-  handleSuccess,
-  handleError,
-}: {
-  accessToken: string;
-  baseUrl: string;
-  osmType: string | undefined;
-  osmId: number | undefined;
-  tagName: string;
-  newTagValue: string;
-  currentOSMObjectOnServer: OSMAPIElement | undefined;
-  handleSuccess: () => void;
-  handleError: () => void;
-}) {
-  const { baseUrl: inhouseBaseUrl } = useInhouseOSMAPI({ cached: false });
-
-  return React.useCallback(async () => {
-    if (
-      !currentOSMObjectOnServer ||
-      !accessToken ||
-      !newTagValue ||
-      !osmType ||
-      !osmId
-    ) {
-      handleError();
-      throw new Error(
-        "Some information was missing while saving to OpenStreetMap. Please let us know if the error persists.",
-      );
-    }
-
-    let changesetComplete = false;
-
-    try {
-      const changesetId = await createChangeset({
-        baseUrl,
-        accessToken,
-        tagName,
-        newValue: newTagValue,
-      });
-
-      await createChange({
-        baseUrl,
-        accessToken,
-        osmType,
-        osmId,
-        changesetId,
-        tagName,
-        newTagValue,
-        currentOSMObjectOnServer,
-      });
-
-      changesetComplete = true;
-
-      // user gets a success toast as soon as the changes are successfully made on osm
-      handleSuccess();
-
-      await callBackendToUpdateInhouseDb({
-        baseUrl: inhouseBaseUrl,
-        osmType,
-        osmId,
-        tagName,
-      });
-    } catch (error) {
-      if (changesetComplete) {
-        // log
-      } else {
-        // log
-      }
-    }
-  }, [
-    inhouseBaseUrl,
-    baseUrl,
-    currentOSMObjectOnServer,
-    accessToken,
-    newTagValue,
-    tagName,
-    osmId,
-    osmType,
-    handleSuccess,
-    handleError,
-  ]);
-}
